+++ client/src/components/debug.page.vue
@@ -5,10 +5,20 @@ import Tabs from './../components/debugger/Tabs.vue';
 import { api } from '@/api';
 import { onMounted, ref } from 'vue';
 import type { DebuggerResponse, Session } from '@/api/types/debugger';
+import { DebuggerState } from '@/api/types/debugger';
 
 const sessions = ref<Session[]>([]);
 const activeSession = ref<string | null>(null);
 const debuggerResponse = ref<DebuggerResponse | null>(null);
+
+const isActiveSessionRunning = () => {
+  const session = sessions.value.find(s => s.id === activeSession.value);
+  if (!session) {
+    return false;
+  }
+
+  return session.state === DebuggerState.Running;
+}
 
 async function toggleActiveSession(sessionIdSelected: string) {
   activeSession.value =
@@ -49,7 +59,7 @@ onMounted(async () => {
 <template>
   <div class="container is-fluid" style="height: calc(100vh - 60px)">
     <div style="height: 150px">
-      <Toolbar :is-active="!!activeSession"/>
+      <Toolbar :is-active="isActiveSessionRunning()"/>
       <Tabs
           :sessions="sessions"
           :active-session="activeSession"