+++ client/src/components/ui/code.vue
@@ -1,0 +1,13 @@
+<script setup lang="ts">
+
+defineProps<{ code: string }>();
+
+const nl2br = (v: string): string => {
+  const res = v.replace(/\r?\n/g, '<br/>');
+  return res;
+}
+</script>
+
+<template>
+  <pre class="code">{{ code }}</pre>
+</template>