+++ src/redis-client/resp/converter.ts
@@ -2,6 +2,7 @@ import { RespEncoder } from './encoder';
 import { RespDecoder } from './decoder';
 import { RedisValue } from './types';
 import { PayloadExtractor } from './payload/extractor';
+import { Renderer } from './renderer';
 
 /**
  * https://redis.io/docs/reference/protocol-spec/
@@ -9,10 +10,12 @@ import { PayloadExtractor } from './payload/extractor';
 export class RespConverter {
   private decoder: RespDecoder;
   private encoder: RespEncoder;
+  private renderer: Renderer;
 
   constructor() {
     this.encoder = new RespEncoder();
     this.decoder = new RespDecoder();
+    this.renderer = new Renderer();
   }
 
   encodeRequest(request: string[]): string {
@@ -23,13 +26,17 @@ export class RespConverter {
     return this.encoder.encode(value);
   }
 
+  render(value: RedisValue): string {
+    return this.renderer.render(value);
+  }
+
   decode(payload: string): RedisValue {
     const extractor = new PayloadExtractor(payload);
 
     return this.decoder.decode(extractor);
   }
 
-  decodeRequest(payload: string): RedisValue[] {
+  decodeRequest(payload: string): string[][] {
     return this.decoder.decodeRequest(payload);
   }